+++ common/src/main/java/com/nemonotfound/nemos/inventory/sorting/client/gui/components/SortAlphabeticallyButton.java
@@ -1,7 +1,5 @@
 package com.nemonotfound.nemos.inventory.sorting.client.gui.components;
 
-import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
-import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 
 import static com.nemonotfound.nemos.inventory.sorting.Constants.MOD_ID;
@@ -11,8 +9,8 @@ public class SortAlphabeticallyButton extends AbstractSortAlphabeticallyButton {
     private final ResourceLocation buttonTexture = ResourceLocation.fromNamespaceAndPath(MOD_ID, "sort_button_alphabetically_inc");
     private final ResourceLocation buttonHoverTexture = ResourceLocation.fromNamespaceAndPath(MOD_ID, "sort_button_alphabetically_inc_highlighted");
 
-    public SortAlphabeticallyButton(int x, int y, int xOffset, int width, int height, Component component, AbstractContainerScreen<?> containerScreen) {
-        super(x, y, xOffset, width, height, component, containerScreen);
+    public SortAlphabeticallyButton(Builder<SortAlphabeticallyButton> builder) {
+        super(builder);
     }
 
     @Override