+++ common/src/main/java/com/nemonotfound/nemos/inventory/sorting/client/gui/components/SortAlphabeticallyDescendingButton.java
@@ -1,7 +1,5 @@
 package com.nemonotfound.nemos.inventory.sorting.client.gui.components;
 
-import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
-import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
 
@@ -15,8 +13,8 @@ public class SortAlphabeticallyDescendingButton extends AbstractSortAlphabetical
     private final ResourceLocation buttonTexture = ResourceLocation.fromNamespaceAndPath(MOD_ID, "sort_button_alphabetically_dec");
     private final ResourceLocation buttonHoverTexture = ResourceLocation.fromNamespaceAndPath(MOD_ID, "sort_button_alphabetically_dec_highlighted");
 
-    public SortAlphabeticallyDescendingButton(int x, int y, int xOffset, int width, int height, Component component, AbstractContainerScreen<?> containerScreen) {
-        super(x, y, xOffset, width, height, component, containerScreen);
+    public SortAlphabeticallyDescendingButton(Builder<SortAlphabeticallyDescendingButton> builder) {
+        super(builder);
     }
 
     @Override