+++ Changelog.md
@@ -1,7 +1,12 @@
 # Changelog v1.2
 
 ## Additions
+- Added drop all button
 - Added tooltip to buttons
+- Sort buttons for the inventory now shows also when in container menu (e.g. chest)
+
+## Changes
+- Resized buttons
 
 ## Fixes
 - Button still exists in creative mode tab and sorts the content